pyproject.toml
[project]
name = "rag-test"
version = "0.1.0"
description = "Agentic RAG with LangGraph + DeepSeek + Self-Reflection"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Kang", email = "zhoukangchn@example.com" }
]
keywords = ["rag", "agent", "langgraph", "fastapi", "deepseek"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "tavily-python>=0.7.20",
    "uvicorn>=0.40.0",
    "prometheus-api-client>=0.5.5",
    "kr8s>=0.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
rag-agent = "src.api.app:app"

[project.urls]
Homepage = "https://github.com/zhoukangchn/rag-test"
Repository = "https://github.com/zhoukangchn/rag-test"
Issues = "https://github.com/zhoukangchn/rag-test/issues"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Ruff 配置
[tool.ruff]
target-version = "py311"
line-length = 100
exclude = [".venv", "__pycache__", ".git", "app"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# Pytest 配置
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

# MyPy 配置
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

# Coverage 配置
[tool.coverage.run]
source = ["src"]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
